# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AccommodationOption {
  slug: string;
  name: string;
  availableRooms: number;
  maxGuests: number;
  description: string;
  pricePerNight: number;
  totalPrice: number;
  photos: string[];
  amenities: string[];
}

export interface AvailabilityResponse {
  available: boolean;
  roomsAvailable: number;
  /** @nullable */
  accommodationSlug: string | null;
  /** @nullable */
  accommodationName: string | null;
  message: string;
  options: AccommodationOption[];
}

export interface ReservationInput {
  checkin: string;
  checkout: string;
  /**
     * @minimum 1
     * @maximum 6
     */
  guests: number;
  /**
     * @minimum 1
     * @maximum 3
     */
  rooms: number;
  accommodationSlug: string;
  /** @nullable */
  offerCode?: string | null;
  /** @minLength 2 */
  guestName: string;
  guestEmail: string;
  /** @minLength 8 */
  guestPhone: string;
  /** @nullable */
  specialRequests?: string | null;
}

export type ReservationStatus = typeof ReservationStatus[keyof typeof ReservationStatus];


export const ReservationStatus = {
  confirmed: 'confirmed',
  cancelled: 'cancelled',
} as const;

export interface Reservation {
  id: number;
  confirmationCode: string;
  status: ReservationStatus;
  checkin: string;
  checkout: string;
  guests: number;
  rooms: number;
  totalAmount: number;
  accommodationSlug: string;
  accommodationName: string;
  guestName: string;
  guestEmail: string;
  guestPhone: string;
  /** @nullable */
  offerCode: string | null;
  /** @nullable */
  offerName: string | null;
  createdAt: string;
}

export type ReservationUpdateStatus = typeof ReservationUpdateStatus[keyof typeof ReservationUpdateStatus];


export const ReservationUpdateStatus = {
  confirmed: 'confirmed',
  cancelled: 'cancelled',
} as const;

export interface ReservationUpdate {
  status?: ReservationUpdateStatus;
  checkin?: string;
  checkout?: string;
  /**
     * @minimum 1
     * @maximum 6
     */
  guests?: number;
  /**
     * @minimum 1
     * @maximum 3
     */
  rooms?: number;
  accommodationSlug?: string;
  /** @nullable */
  offerCode?: string | null;
}

export type CheckinParameter = string;

export type CheckoutParameter = string;

export type GuestsParameter = number;

export type RoomsParameter = number;

export type CheckAvailabilityParams = {
checkin: CheckinParameter;
checkout: CheckoutParameter;
/**
 * @minimum 1
 * @maximum 6
 */
guests: GuestsParameter;
/**
 * @minimum 1
 * @maximum 3
 */
rooms: RoomsParameter;
};