# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: reservations
    description: Hotel availability and reservation operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /rooms/availability:
    get:
      operationId: checkAvailability
      tags: [reservations]
      summary: Check hotel availability
      description: Checks room availability for a date range and guest count.
      parameters:
        - $ref: "#/components/parameters/Checkin"
        - $ref: "#/components/parameters/Checkout"
        - $ref: "#/components/parameters/Guests"
        - $ref: "#/components/parameters/Rooms"
      responses:
        "200":
          description: Availability result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityResponse"
        "400":
          description: Invalid dates or guest counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reservations:
    post:
      operationId: createReservation
      tags: [reservations]
      summary: Create a hotel reservation
      description: Creates a confirmed reservation after checking for overlapping bookings.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationInput"
      responses:
        "201":
          description: Confirmed reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "400":
          description: Invalid reservation details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: No rooms available for the requested stay
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reservations/{id}:
    get:
      operationId: getReservation
      tags: [reservations]
      summary: Get a reservation
      parameters:
        - $ref: "#/components/parameters/ReservationId"
      responses:
        "200":
          description: Reservation details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "404":
          description: Reservation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateReservation
      tags: [reservations]
      summary: Cancel or update a reservation
      parameters:
        - $ref: "#/components/parameters/ReservationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationUpdate"
      responses:
        "200":
          description: Updated reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "400":
          description: Invalid reservation update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Reservation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Updated dates are unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  parameters:
    Checkin:
      name: checkin
      in: query
      required: true
      schema:
        type: string
        format: date
    Checkout:
      name: checkout
      in: query
      required: true
      schema:
        type: string
        format: date
    Guests:
      name: guests
      in: query
      required: true
      schema:
        type: number
        minimum: 1
        maximum: 6
    Rooms:
      name: rooms
      in: query
      required: true
      schema:
        type: number
        minimum: 1
        maximum: 3
    ReservationId:
      name: id
      in: path
      required: true
      schema:
        type: number
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    AccommodationOption:
      type: object
      properties:
        slug:
          type: string
        name:
          type: string
        availableRooms:
          type: number
        maxGuests:
          type: number
        description:
          type: string
        pricePerNight:
          type: number
        totalPrice:
          type: number
        photos:
          type: array
          items:
            type: string
        amenities:
          type: array
          items:
            type: string
      required:
        - slug
        - name
        - availableRooms
        - maxGuests
        - description
        - pricePerNight
        - totalPrice
        - photos
        - amenities
    AvailabilityResponse:
      type: object
      properties:
        available:
          type: boolean
        roomsAvailable:
          type: number
        accommodationSlug:
          type: ["string", "null"]
        accommodationName:
          type: ["string", "null"]
        message:
          type: string
        options:
          type: array
          items:
            $ref: "#/components/schemas/AccommodationOption"
      required:
        - available
        - roomsAvailable
        - accommodationSlug
        - accommodationName
        - message
        - options
    ReservationInput:
      type: object
      properties:
        checkin:
          type: string
          format: date
        checkout:
          type: string
          format: date
        guests:
          type: number
          minimum: 1
          maximum: 6
        rooms:
          type: number
          minimum: 1
          maximum: 3
        accommodationSlug:
          type: string
        offerCode:
          type: ["string", "null"]
        guestName:
          type: string
          minLength: 2
        guestEmail:
          type: string
        guestPhone:
          type: string
          minLength: 8
        specialRequests:
          type: ["string", "null"]
      required:
        - checkin
        - checkout
        - guests
        - rooms
        - accommodationSlug
        - guestName
        - guestEmail
        - guestPhone
    Reservation:
      type: object
      properties:
        id:
          type: number
        confirmationCode:
          type: string
        status:
          type: string
          enum: [confirmed, cancelled]
        checkin:
          type: string
          format: date
        checkout:
          type: string
          format: date
        guests:
          type: number
        rooms:
          type: number
        totalAmount:
          type: number
        accommodationSlug:
          type: string
        accommodationName:
          type: string
        guestName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: string
        offerCode:
          type: ["string", "null"]
        offerName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - confirmationCode
        - status
        - checkin
        - checkout
        - guests
        - rooms
        - totalAmount
        - accommodationSlug
        - accommodationName
        - guestName
        - guestEmail
        - guestPhone
        - offerCode
        - offerName
        - createdAt
    ReservationUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [confirmed, cancelled]
        checkin:
          type: string
          format: date
        checkout:
          type: string
          format: date
        guests:
          type: number
          minimum: 1
          maximum: 6
        rooms:
          type: number
          minimum: 1
          maximum: 3
        accommodationSlug:
          type: string
        offerCode:
          type: ["string", "null"]