# lib/api-zod/src/generated/types/accommodationOption.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AccommodationOption {
  slug: string;
  name: string;
  availableRooms: number;
  maxGuests: number;
  description: string;
  pricePerNight: number;
  totalPrice: number;
  photos: string[];
  amenities: string[];
}