# lib/api-zod/src/generated/types/reservation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReservationStatus } from './reservationStatus';

export interface Reservation {
  id: number;
  confirmationCode: string;
  status: ReservationStatus;
  checkin: Date;
  checkout: Date;
  guests: number;
  rooms: number;
  totalAmount: number;
  accommodationSlug: string;
  accommodationName: string;
  guestName: string;
  guestEmail: string;
  guestPhone: string;
  /** @nullable */
  offerCode: string | null;
  /** @nullable */
  offerName: string | null;
  createdAt: Date;
}