# lib/api-zod/src/generated/types/reservationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReservationInput {
  checkin: Date;
  checkout: Date;
  /**
     * @minimum 1
     * @maximum 6
     */
  guests: number;
  /**
     * @minimum 1
     * @maximum 3
     */
  rooms: number;
  accommodationSlug: string;
  /** @nullable */
  offerCode?: string | null;
  /** @minLength 2 */
  guestName: string;
  guestEmail: string;
  /** @minLength 8 */
  guestPhone: string;
  /** @nullable */
  specialRequests?: string | null;
}