# lib/api-zod/src/generated/types/reservationUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReservationUpdateStatus } from './reservationUpdateStatus';

export interface ReservationUpdate {
  status?: ReservationUpdateStatus;
  checkin?: Date;
  checkout?: Date;
  /**
     * @minimum 1
     * @maximum 6
     */
  guests?: number;
  /**
     * @minimum 1
     * @maximum 3
     */
  rooms?: number;
  accommodationSlug?: string;
  /** @nullable */
  offerCode?: string | null;
}